LayoutTests/platform/win/media/W3C/video/canPlayType/canPlayType_two_implies_one_6-expected.txt

video.canPlayType() - codecs parameter logic

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL "probably" should be . Was probably.

TEST COMPLETE
spec reference
